Add expandBoundingBoxFactor method to stage
This can be useful if NGL is embedded and additional graphics objects
are added to the stage outside NGL. Without expanding the bounding
box, those additional objects may be clipped or fogged.

This is the improved version of this patch, with better math
and also it only applies the expansion once.