Ensure the OES_standard_derivatives extension is enabled for WebGL1 browsers.
Three.js r112 introduced a requirement for OES_standard_derivatives
when using the lights_physical_fragment shader chunk.

In WebGL2 this extension is automatically enabled / builtin.

The extension is now enabled in all shaders created through the Buffer
class - if this impacts performance we could enable/disable as
actually required, or change the lighting model.