WIP: use three.js color management for color accuracy
Set the renderer output to the recommended sRGB encoding,
and linearize all object colors so the internal three.js
workflow is fully linear. This also allows correct glTF
export of models.

This is WIP because it doesn't propagate the color management
settings from the top-level viewer down to the colormaker yet.