Add selection-centered rotation & pivot point
This adds a `pivot` property to Component, with a `setPivot` method.

It also arranges for StructureComponents to optionally pass the
current selection to Component.updateMatrix when it gets the
component's center.

StructureComponents now can set their centering mode to "selection" or
"component". This makes components able to rotate around their
selection's center, not just the complete structure's center.

To change the centering mode, call new method
`StructureComponent.setCenterMode()`. The default is set to
"selection", which is new & improved behavior. Set it to "component"
to get the old behavior.

This also adds a pivot.js example to demonstrate the pivot feature in
a couple of different ways, as well as demonstrate selection-based
centering, showing how to update the transform when changing selection
to avoid molecules jumping around. (We could do that in the core code,
but since it uses Component.transform, users who are already using
that would be impacted -- so probably best to just show it in an
example and users can adapt it to suit.)