Set default surface type to 'av', increase min scaleFactor
Per comments in #843, use `av` as the default surface type rather than
`vws`. It looks better, and as Fred Ludlow points out the 'av' surface
is also a Connolly surface which has the nice property that protein
and ligand surfaces generally 'mesh' quite nicely at a probe-size of
about 1.4-1.5, so for visual inspection of how well a ligand fits a
binding site it's a good choice.

Also For molecules like 1D3E that trigger `backboneOnly`, the
scaleFactor (isosurface volume resolution) was getting reduced to
0.15, which is really small and gives a very "low-poly" look. This
sets it to 0.5 which should still be fast enough for slow machines but
looks much better.