Update Typescript from 3.8.3 to 4.5.4
This also pulls in some other more recent dependencies, like eslint
and jest.

Typescript 4 is stricter in some cases, so there are a few changes to
the code to reflect that. The only significant ones are that
typescript 4 correctly doesn't let you delete a non-optional field of
an object -- that would leave the object in a broken state. Instgead
of deleting things like a structureRepresentation's structure, we just
dispose of it or set it to an empty object or array. That will have
the same effect; the unlinked object can be garbage collected.