fix CL1 not recognized as CL in parm7 files
This issue was noticed in parm7 file (Amber topology file),
where atoms labelled CL1 were identified as Carbon instead of Chlorine.
The `guessElement` method was designed to strip up to 2 digits on the
left side of the atom name.
This was leading CL1 to be considered as a 3 letters atomname and processed
using the first letter only.

Given that the former version of the function was designed
to execute a limited number
of fast to execute operations
(probably because it is a "hot" function), this commit intends
to extract the first group of letters
while not calling regexps nor parseInt.