modelStore chain counts consistency
In some cases the original number of chains could have been greater
than the number of chains after the calculateChainnames execution.
As the chainCount in the modelStore was not reset, it was possible
to execute the model.eachChain() callback more times than the
actual number of chain, over references that were non longer valid.
A side effect of this was the creation of PDB files that contained
duplicated sections.

This fix ensures that the chainCount value in the modelStore is consistent
with the execution of the `addChain` function.