#999 _chem_comp can be defined on mmCIF files
The code was assuming that the _chem_comp property was only available
for non macromolecular structures (such as cif
files for hetero atoms entities).
It appears that some PDB files (e.g. `32C2`) may define
a _chem_comp loop around every
chemical component such as its amino-acids.
This breaks the previous code as `cif.chem_comp` then contains
an array of strings instead of a string.
The fix here is to assume that if `struct` is present then we should not
enter in the block for parsing simple entities.