convert color when appending to uniforms array
ColorMakers are called when creating BufferArrays to populate the `color` uniform.
4 methods can be applied (atomColorToArray, bondColorToArray, volumeColorToArray, positionColorToArray).
Each of these, calls a related method (e.g. `atomColor`) to get a color that is defined in RGB Colorspace, encoded as an integer between 0 and 2**24.
This color is then converted to a tuple of 3 values which are passed along all the other vertices colors to the shaders codes as a uniform.
Prior code was using a decorator to convert the output color of each coloring function, to linear space and then to hex.
This hex value was then later processed to be split and normalized in the array buffer.
The change here makes the linearization at the last step, after the normalization which should avoid clamping due to the double conversion that was happening previously