Get connectivity from chem_comp_bond
This table is available in "Updated mmcif files" distributed by PDBe.
In this commit, the list of bonds defined for each residue is stored in a new dictionary in a ChemCompMap object.
Bonds in the mmcif file are defined using atom names (e.g. CA), which need to be converted in indices in the atomList from a given residue type.
The atomList contain list of indices of AtomTypes from the structure AtomMap.
Given an atom index from the atomAlist, the AtomMap.get(idx) method returns an AtomType object that contains the atomname property.