fix: jest cannot import modules from molstar dir
Jest cannot import code from ES modules which is the case of the modules from MolStar (not bundled).
Jest code fails with some indications about tweaking jest config using the transformIgnorePattern property. After much trials and research I was not able to make it work and decided to switch the test runner to vitest, which solved the issue.